Moving test from global and putting it in cursor
