+++ tests/test_001_globals.py
@@ -1059,28 +1059,3 @@ def test_row_string_key_case_insensitive_with_lowercase():
     # Non-existent attribute raises AttributeError
     with pytest.raises(AttributeError):
         row.nonexistent
-
-
-def test_map_sql_type_none_returns_sql_varchar():
-    """Test that _map_sql_type returns SQL_VARCHAR for None params (GH-610).
-
-    Previously, None returned SQL_UNKNOWN_TYPE which triggered a costly
-    SQLDescribeParam / sp_describe_undeclared_parameters round-trip to the
-    server on every execute call with a NULL parameter.
-    """
-    from unittest.mock import MagicMock
-
-    from mssql_python.constants import ConstantsDDBC as ddbc_sql_const
-
-    cursor = MagicMock(spec=mssql_python.Cursor)
-    # Bind the real method to the mock so we test actual logic
-    _map_sql_type = mssql_python.Cursor._map_sql_type.__get__(cursor)
-    params = [None, 42, None]
-
-    sql_type, c_type, col_size, dec_digits, is_dae = _map_sql_type(None, params, 0)
-
-    assert sql_type == ddbc_sql_const.SQL_VARCHAR.value
-    assert c_type == ddbc_sql_const.SQL_C_DEFAULT.value
-    assert col_size == 1
-    assert dec_digits == 0
-    assert is_dae is False
